\name{Auditory_clusterTH3_2}
\alias{Auditory_clusterTH3_2}
\docType{data}
\title{
Auditory cluster threshold 3.2
}
\description{
Image in NIfTI format describing the cluster map computed by FSL considering a threshold equals 3.2
}
\usage{data("Auditory_clusterTH3_2")}
\format{
  The format is:
 'niftiImage' int [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
 - attr(*, "pixdim")= num [1:3] 2 2 2
 - attr(*, "pixunits")= chr [1:2] "mm" "s"
 - attr(*, ".nifti_image_ptr")=<externalptr>
}

\examples{
data(Auditory_clusterTH3_2)
## maybe str(Auditory_clusterTH3_2) ; plot(Auditory_clusterTH3_2) ...
}
\keyword{datasets}
