\name{Auditory_copes}
\alias{Auditory_copes}
\docType{data}
\title{
Auditory copes
}
\description{
List of copes in NifTI format
}
\usage{data("Auditory_copes")}
\format{
  The format is:
List of 140
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] -2.44 -2.44 -2.44 -2.44 -2.44 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
  [list output truncated]
}
\source{
https://openneuro.org/datasets/ds000158/versions/1.0.0
}
\details{
List of copes from FLAME 1 analysis computed by (FSL)[https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FEAT/UserGuide#Single-Group_Average_.28One-Sample_T-Test.29]
}
\examples{
data(Auditory_copes)
str(Auditory_copes)
}
\keyword{datasets}
