% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRIdata.R
\docType{package}
\name{fMRIdata}
\alias{fMRIdata}
\title{functional Magnetic Resonance Imaging (fMRI) Data Package}
\description{
This package is meant to serve as a data collection tool for fMRI data.
}
\examples{
# Please see the datasets
}
\references{
Please see <https://openneuro.org/>
}
\author{
Angela Andreella \email{angela.andreella@unipd.it},
}
