\name{Food_copes}
\alias{Food_copes}
\docType{data}
\title{
Food copes
}
\description{
List of copes in NifTI format
}
\usage{data("Food_copes")}
\format{
  The format is:
List of 29
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
 $ : 'niftiImage' num [1:91, 1:109, 1:91] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "pixdim")= num [1:3] 2 2 2
  ..- attr(*, "pixunits")= chr [1:2] "mm" "s"
  ..- attr(*, ".nifti_image_ptr")=<externalptr>
}
\source{
https://openneuro.org/datasets/ds000157/versions/00001
}
\details{
List of copes from FLAME 1 analysis computed by (FSL)[https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FEAT/UserGuide#Single-Group_Average_.28One-Sample_T-Test.29]
}
\examples{
data(Food_copes)
str(Food_copes)
}
\keyword{datasets}
