\name{WO_copes}
\alias{WO_copes}
\docType{data}
\title{
Word Object copes
}
\description{
List of copes in NifTI format
}
\usage{data("WO_copes")}
\format{
  The format is:
 list()
}
\details{
List of copes from FLAME 1 analysis computed by (FSL)[https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FEAT/UserGuide#Single-Group_Average_.28One-Sample_T-Test.29]
}
\source{
https://openneuro.org/datasets/ds000107/versions/00001
}

\examples{
data(WO_copes)
str(WO_copes)
}
\keyword{datasets}
